pyproject.toml
[project]
name = "text-adv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "art>=6.4",
    "google-genai>=1.7.0",
    "google-generativeai>=0.8.4",
    "markdown-it-py==3.0.0",
    "mdurl==0.1.2",
    "numpy>=2.2.4",
    "overrides>=7.7.0",
    "pygments==2.19.1",
    "python-dotenv>=1.0.1",
    "pyyaml==6.0.2",
    "readchar==4.2.1",
    "rich==13.9.4",
    "rich-pixels>=3.0.1",
    "ruff>=0.11.0",
    "six==1.17.0",
    "sounddevice>=0.5.1",
    "soundfile>=0.13.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
